+++ EspionSpotify/Enums/SilenceAnalyzer.cs
@@ -1,0 +1,9 @@
+﻿namespace EspionSpotify.Enums
+{
+    public enum SilenceAnalyzer
+    {  
+        None,
+        TrimEnd,
+        TrimStart
+    }
+}
