+++ EspionSpotify/Models/AudioWaveBuffer.cs
@@ -1,0 +1,9 @@
+﻿namespace EspionSpotify.Models
+{
+    public class AudioWaveBuffer
+    {
+        public byte[] Buffer;
+        public int BytesRecordedCount;
+        public bool WithSilence;
+    }
+}
